Work-around for ndarray dtype object segfault
There seems to be a bug a in numpy that causes a segfault when restoring
a binary-encoded ndarray of dtype object. E.g., this will segfault:

    import numpy as np
    arr = np.array([str(i) for i in range(3)], dtype=np.object)
    dtype = arr.dtype
    shape = arr.shape
    buf = arr.tobytes()
    del arr
    arr = np.ndarray(buffer=buf, dtype=dtype, shape=shape).copy()

Removing the 'del arr' avoids the segfault, so it seems to be a
low-level pointer bug in numpy.ndarray. Other dtypes that I've tested
work fine, so we special-case object dtypes and save those as a
json-encoded list rather than a byte array.